Add extra case for mongodb

+++ test/check_registry/check_registry.sh
@@ -84,12 +84,17 @@ do
 
     # If the correct devfile is not matched throw error
     if [ "$found_matching" == "0" ]; then
-        let ENTRIES_FAILED++
-        echo "[FAIL] Failed to match project $repo with expected $devfile. Command ./alizer devfile --registry $registry $path"
-        echo "[FAIL] Output: $alizer_output"
-        echo "(PASSED: $ENTRIES_PASSED / FAILED: $ENTRIES_FAILED)"
-        rm -rf tmp/$devfile
-        exit 1
+        # nodejs-mongodb similar to nodejs
+        if [[ "$repo" == "https://github.com/che-samples/nodejs-mongodb-sample" && "$devfile" == "nodejs-mongodb" ]]; then
+            echo "[WARNING] $repo matched with $devfile. Will not raise error."
+        else
+            let ENTRIES_FAILED++
+            echo "[FAIL] Failed to match project $repo with expected $devfile. Command ./alizer devfile --registry $registry $path"
+            echo "[FAIL] Output: $alizer_output"
+            echo "(PASSED: $ENTRIES_PASSED / FAILED: $ENTRIES_FAILED)"
+            rm -rf tmp/$devfile
+            exit 1
+        fi
     fi
     rm -rf tmp/$devfile
 done
